third_party/blink/web_tests/http/tests/webfont/webfont-performance-duration-expected.txt

PASS entry.startTime >= waitMillisBeforeFetching is true
PASS successfullyParsed is true

TEST COMPLETE

Hello webfont
